pyproject.toml
[project]
name = "impactscope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.45",
    "matplotlib>=3.10.8",
    "networkx>=3.4.2",
    "pre-commit>=4.5.0",
    "pyvis>=0.3.2",
    "tree-sitter>=0.25.2",
    "tree-sitter-c>=0.24.1",
    "tree-sitter-languages>=1.10.2",
    "typer>=0.20.0",
]
